modularization: favorite data as single module
