+++ common/favorite/.gitignore
