+++ common/favorite/src/main/java/com/my/favorite/domain/usecase/FetchAllFavoriteUseCase.kt
@@ -1,0 +1,9 @@
+package com.my.favorite.domain.usecase
+
+import com.my.core.extensions.applySchedulers
+import com.my.favorite.domain.repository.FavoriteRepository
+import io.reactivex.Flowable
+
+class FetchAllFavoriteUseCase(private val repository: FavoriteRepository) {
+    operator fun invoke(): Flowable<List<FavoriteEntity>> = repository.loadAll().applySchedulers()
+}
