--- shared/data/movies/src/main/java/com/my/movie/favorite/Converter.kt
@@ -1,22 +1,0 @@
-package com.my.movie.favorite
-
-import com.my.domain.entity.MovieDetails
-import com.my.movie.favorite.dto.FavoriteEntity
-
-internal fun MovieDetails.toFavorite(): FavoriteEntity = FavoriteEntity(
-    id = this.id,
-    title = this.title,
-    posterPath = this.posterPath,
-    rating = this.rating
-)
-
-internal fun FavoriteEntity.toMovieDetails() = MovieDetails(
-    id = this.id,
-    title = this.title,
-    overview = "this.overview",
-    posterPath = this.posterPath,
-    rating = this.rating,
-    studio = "",
-    genre = "",
-    year = ""
-)
