--- shared/data/movies/src/main/java/com/my/movie/favorite/dto/FavoriteEntity.kt
@@ -1,13 +1,0 @@
-package com.my.movie.favorite.dto
-
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-
-@Entity
-internal data class FavoriteEntity(
-    @PrimaryKey
-    val id: Int,
-    val title: String,
-    val posterPath: String,
-    val rating: Float
-)
