--- shared/domain/src/main/java/com/my/domain/usecase/DeleteFromFavoriteByIdUseCase.kt
@@ -1,9 +1,0 @@
-package com.my.domain.usecase
-
-import com.my.domain.repository.FavoriteRepository
-import com.my.resources.extensions.applySchedulers
-import io.reactivex.Completable
-
-class DeleteFromFavoriteByIdUseCase(private val repository: FavoriteRepository) {
-    operator fun invoke(id: Int): Completable = repository.deleteById(id).applySchedulers()
-}
