--- shared/domain/src/main/java/com/my/domain/usecase/FetchAllFavoriteUseCase.kt
@@ -1,10 +1,0 @@
-package com.my.domain.usecase
-
-import com.my.domain.entity.MovieDetails
-import com.my.domain.repository.FavoriteRepository
-import com.my.resources.extensions.applySchedulers
-import io.reactivex.Flowable
-
-class FetchAllFavoriteUseCase(private val repository: FavoriteRepository) {
-    operator fun invoke(): Flowable<List<MovieDetails>> = repository.loadAll().applySchedulers()
-}
